requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
pydantic
httpx
python-dotenv
python-multipart
pypresence
requests
